Refactored to add Comment and AddComment components
