+++ frontend/src/core/translation/components/AddComment.js
@@ -1,0 +1,14 @@
+/* @flow */
+
+import * as React from 'react';
+
+type Props = {|
+    
+|};
+
+
+export default class Comments extends React.Component<Props> {
+    render() {
+        return null;
+    }
+}
