+++ frontend/src/core/translation/components/Comments.js
@@ -1,88 +1,31 @@
 /* @flow */
 
 import * as React from 'react';
-import ReactTimeAgo from 'react-time-ago';
 
-import { WithPlaceables } from 'core/placeable';
 import type { TranslationComment } from 'core/api';
+
+import { Comment } from 'core/translation';
 
 
 type Props = {|
-    comment: TranslationComment,
+    translationComments: Array<TranslationComment>,
 |};
 
 
 export default class Comments extends React.Component<Props> {
-    renderUserAvatar() {
-        const { comment } = this.props;
-    
-        if (!comment.author_id) {
-            return <img
-                src='/static/img/anon.jpg'
-                alt=''
-                height='44'
-                width='44'
-            />;
-        }
-
-        return <a
-            href={ `/contributors/${comment.username}` }
-            title='Author'
-            target='_blank'
-            rel='noopener noreferrer'
-            onClick={ (e: SyntheticMouseEvent<>) => e.stopPropagation() }
-        >
-            <img
-                src={ comment.user_gravatar_url_small }
-                alt=''
-                height='44'
-                width='44'
-            />
-        </a>
-    }
-
-    renderUser() {
-        const { comment } = this.props;
-
-        if (!comment.author_id) {
-            return <span>{ comment.author }</span>;
-        }
-
-        return <a
-            href={ `/contributors/${comment.username}` }
-            title='Author'
-            target='_blank'
-            rel='noopener noreferrer'
-            onClick={ (e: SyntheticMouseEvent<>) => e.stopPropagation() }
-        >
-            { comment.username }
-        </a>
-    }
-
     render() {
-        const { comment } = this.props;
+        const { translationComments } = this.props;
 
-        if (!comment) {
+        if (!translationComments) {
             return null;
         }
 
-        var divStyle = { display: 'inline' }
-
-        return <div>
-                <li title='Comment'>
-                    <div className='avatar' style={divStyle}>
-                        { this.renderUserAvatar() }
-                    </div>
-                    { this.renderUser() }
-                    <WithPlaceables> { comment.content } </WithPlaceables>
-                </li>
-                <li>
-                    <ReactTimeAgo
-                        dir='ltr'
-                        date={ new Date(comment.date_iso) }
-                        title={ `${comment.timestamp} UTC` }
-                    />
-                </li>
-            </div>
+        return (
+            <li>
+                { translationComments.map(content =>
+                    <Comment comment={ content } />
+                )}
+            </li>
+        )
     }
 }
