+++ frontend/src/core/translation/index.js
@@ -3,3 +3,5 @@
 export { default as GenericTranslation } from './components/GenericTranslation';
 export { default as TranslationProxy } from './components/TranslationProxy';
 export { default as Comments } from './components/Comments';
+export { default as Comment } from './components/Comment';
+export { default as AddComment } from './components/AddComment';
